refactor(tt): extract book chapter components
