+++ apps/total-typescript/src/components/book/use-visible-markdown-heading.tsx
@@ -1,0 +1,49 @@
+import React from 'react'
+
+type UseVisibleMarkdownHeadingOptions = {
+  rootMargin?: string
+  threshold?: number | number[]
+}
+
+export const useVisibleMarkdownHeading = (
+  headingIds: string[],
+  options?: UseVisibleMarkdownHeadingOptions,
+): string | null => {
+  const [visibleHeadingId, setVisibleHeadingId] = React.useState<string | null>(
+    null,
+  )
+
+  React.useEffect(() => {
+    const observer = new IntersectionObserver(
+      (entries) => {
+        entries.forEach((entry) => {
+          if (entry.isIntersecting) {
+            setVisibleHeadingId(entry.target.id)
+          }
+        })
+      },
+      {
+        rootMargin: options?.rootMargin || '0px',
+        threshold: options?.threshold || 0,
+      },
+    )
+
+    headingIds.forEach((id) => {
+      const element = document.getElementById(id)
+      if (element) {
+        observer.observe(element)
+      }
+    })
+
+    return () => {
+      headingIds.forEach((id) => {
+        const element = document.getElementById(id)
+        if (element) {
+          observer.unobserve(element)
+        }
+      })
+    }
+  }, [headingIds, options])
+
+  return visibleHeadingId
+}
